Documentation
Added summaries and descriptions to methods and objects
