+++ Blocks/Blocks.csproj
@@ -44,6 +44,7 @@
     <Reference Include="System.Windows.Forms" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Functions\Helpers.cs" />
     <Compile Include="Objects\BlockDefinition.cs" />
     <Compile Include="Objects\BlockInstance.cs" />
     <Compile Include="Objects\BlockAssembly.cs" />
